requirements.txt
# Core voice assistant dependencies
pvporcupine==3.0.2
pyaudio==0.2.14
numpy>=1.24.0
scipy>=1.10.0

# Speech-to-Text (Whisper)
torch>=2.0.0
transformers>=4.35.0
librosa>=0.10.0
soundfile>=0.12.0
openai-whisper>=20231117

# Large Language Model (Llama)  
# Note: Install with: CMAKE_ARGS="-DLLAMA_HIPBLAS=on" pip install llama-cpp-python --force-reinstall --no-cache-dir
llama-cpp-python>=0.2.20

# Text-to-Speech
piper-tts>=1.2.0
onnxruntime>=1.16.0
pygame>=2.5.0

# System integration
pyautogui>=0.9.54
psutil>=5.9.0
dbus-python>=1.3.2

# Audio processing
webrtcvad>=2.0.10
pydub>=0.25.1

# Utilities
pyyaml>=6.0
colorlog>=6.7.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
